refactoring objects
