+++ TestOCR/IEntityToFind.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TestOCR
+{
+    interface IEntityToFind
+    {
+        public void CheckValue(string textExtracted);
+    }
+}
